api/requirements.txt
Flask==2.2
Flask-Cors==3.0.10
requests==2.28
asyncio==3.4.3
psycopg2==2.9.5
numpy==1.24.1
geopandas==0.12.2
pyyaml==6.0
aiohttp==3.8.3
geoalchemy2==0.13
gunicorn==20.1
